requirements.txt
# Основные зависимости
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0

# Для 3D-вьювера
tiny-3d-engine>=0.2.0

# Для работы с PLY файлами
plyfile>=0.7.4

# Для пост-обработки (опционально)
open3d>=0.15.0